DesignPattern data structure
A design pattern for our purposes is a simple graph, where the nodes are
JavaClass objects and connections are relationships between classes
(e.g. extends, implements, etc.).

The DesignPattern class has methods for adding nodes and connections, as
well as an "addHierarchy" method which adds an entire class hierarchy at
once.  This is useful because QDox only gives a flat list of "derived
classes", which makes determining the structure of a hierarchy
difficult.

The Connection class is very basic.  It has "from", "to" and "name"
fields.  We'll probably want to extend this so that more complicated
relationships can be represented (e.g. aggregation).